+++ src/app/shared/components/service-stops.component.ts
@@ -1,6 +1,6 @@
 import { AfterViewInit, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 
-import OJP_Legacy from '../../config/ojp-legacy';
+import * as OJP_Next from 'ojp-sdk-next';
 
 export interface LegStopPointData {
   locationText: string,
@@ -16,7 +16,7 @@ export interface LegStopPointData {
   platformAssistanceIconPath: string | null,
   platformAssistanceTooltip: string,
 
-  geoPosition: OJP_Legacy.GeoPosition | null,
+  geoPosition: OJP_Next.GeoPosition | null,
 
   isNotServicedStop: boolean,
 
@@ -57,6 +57,9 @@ export class ServiceStopsComponent implements OnInit, AfterViewInit {
   }
 
   private emitSelectedEndpoints() {
+    if (this.stopPointsData.length < 2) {
+      return;
+    }
     this.onExampleTripLocationsUpdated.emit(this.selectedEndpointsIDx);
   }