+++ src/app/shared/controllers/trip-geo-controller.ts
@@ -1,6 +1,6 @@
 import OJP_Legacy from '../../config/ojp-legacy';
 
-import { OJPMapHelpers } from '../../helpers/ojp-map-helpers';
+import { JourneyService } from '../models/journey-service';
 import { TripLegDrawType, TripLegLineType, TripLegPropertiesEnum } from '../types/map-geometry-types';
 
 interface LinePointData {
@@ -135,28 +135,13 @@ export class TripLegGeoController {
     if (leg.legType === 'TimedLeg') {
       const timedLeg = leg as OJP_Legacy.TripTimedLeg;
 
-      const usedDetailedLine = TripLegGeoController.serviceHasPrecisePolyline(timedLeg.service);
+      const service = JourneyService.initWithOJP_LegacyJourneyService(timedLeg.service);
+      const usedDetailedLine = service.hasPrecisePolyline();
 
       return !usedDetailedLine;
     }
 
     return defaultValue;
-  }
-
-  private static serviceHasPrecisePolyline(service: OJP_Legacy.JourneyService): boolean {
-    if (service.ptMode.isDemandMode) {
-      return true;
-    }
-
-    const ignorePtModes: string[] = [
-      'bus',
-      'tram'
-    ];
-    if (ignorePtModes.indexOf(service.ptMode.ptMode) !== -1) {
-      return false;
-    }
-
-    return true;
   }
 
   private computeBeelineFeature(): GeoJSON.Feature | null {
@@ -248,7 +233,8 @@ export class TripLegGeoController {
     
     if (this.leg.legType === 'TimedLeg') {
       const timedLeg = this.leg as OJP_Legacy.TripTimedLeg;
-      return OJPMapHelpers.computeTimedLegColor(timedLeg);
+      const service = JourneyService.initWithOJP_LegacyJourneyService(timedLeg.service);
+      return service.computeLegColor();
     }
 
     return defaultType;
@@ -402,7 +388,8 @@ export class TripLegGeoController {
   private computeTimedLegGeoJSONFeatures(timedLeg: OJP_Legacy.TripTimedLeg): GeoJSON.Feature[] {
     let features: GeoJSON.Feature[] = [];
 
-    const lineType: TripLegLineType = OJPMapHelpers.computeTimedLegColor(timedLeg);
+    const service = JourneyService.initWithOJP_LegacyJourneyService(timedLeg.service);
+    const lineType: TripLegLineType = service.computeLegColor();
 
     const useDetailedTrack = !this.useBeeLine;
     if (useDetailedTrack) {