+++ CHANGELOG.md
@@ -4,6 +4,13 @@ OJP-Demo URL: https://opentdatach.github.io/ojp-demo-app/
 
 ----
 
+4.June 2025
+- TripInfoRequest - use new SDK - [PR #243](https://github.com/openTdataCH/ojp-demo-app-src/pull/243)
+  - use `ojp-sdk-next` for TripInfoRequest
+  - extracts `JourneyService`, `TripInfoResult`, `StopPointCall` logic from SDK
+  - adds detailed route for TripInfoRequest map polyline
+  - use `src/app/config/app-config.local.ts` for local developlment
+
 15.May 2025
 - TRR XML debug popover - [PR #233](https://github.com/openTdataCH/ojp-demo-app-src/pull/233)